+++ src/event/kqueueEvent.cpp
@@ -1,7 +1,3 @@
-
-#ifndef EVENT_EVENT_HPP
-#define EVENT_EVENT_HPP
-
 #include <unistd.h>
 
 #include <cstring>
@@ -9,47 +5,7 @@
 #include <stdexcept>
 
 #include "event/kqueue/kqueue.hpp"
-
-enum EventKind { kRead = EVFILT_READ, kWrite = EVFILT_WRITE, kEOF = EV_EOF };
-
-struct Event {
-  EventPool *ep;
-  EventKind kind;
-  intptr_t data;
-};
-
-class EventHandler {
- public:
-  virtual int handle(Event e) = 0;
-  virtual int getFd() = 0;
-};
-
-class EventPool {
- public:
-  EventPool(int max_event);
-  ~EventPool();
-
- private:
-  EventPool(const EventPool &);             // = delete;
-  EventPool &operator=(const EventPool &);  // = delete;
-
- public:
-  int close();
-
-  bool ok();
-
-  int addEvent(EventKind kind, EventHandler *eh);
-
-  int dispatchEvent(time_t sec);
-
-  int dispatchEvent(const struct timespec &ts);
-
- private:
-  struct kevent *event_list_;
-  int fd_;
-  bool ok_;
-  int max_event_;
-};
+#include "event/event.hpp"
 
 EventPool::EventPool(int max_event)
     : event_list_(NULL), ok_(false), max_event_(max_event) {
@@ -113,5 +69,3 @@ int EventPool::dispatchEvent(const struct timespec &ts) {
   }
   return events;
 }
-
-#endif  // EVENT_EVENT_HPP
